consumer/requirements.txt
kafka-python
requests
prometheus-client

